backend/requirements.txt
fastapi
uvicorn
faiss-cpu
sentence-transformers
google-generativeai
python-dotenv
